refactor: extract read_parquet_metas_in_parallel.
